# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "laravel-api脚手架",
        "description": "laravel-api脚手架",
        "contact": {
            "email": "darius@matulionis.lt"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://business-card.local",
            "description": "本地环境"
        },
        {
            "url": "http://dev.dev.com",
            "description": "开发环境"
        },
        {
            "url": "http://dev.stage.com",
            "description": "测试环境"
        },
        {
            "url": "",
            "description": "动态host"
        }
    ],
    "paths": {
...
        "/api/v1/wx-applet/decode-phone-number": {
            "post": {
                "tags": [
                    "微信小程序"
                ],
                "summary": "",
                "description": "解码手机号码",
                "operationId": "解码手机号码",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "encryptedData",
                                    "iv"
                                ],
                                "properties": {
                                    "encryptedData": {
                                        "description": "包括敏感数据在内的完整用户信息的加密数据",
                                        "type": "string"
                                    },
                                    "iv": {
                                        "description": "加密算法的初始向量",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "400": {
                        "$ref": "#/components/responses/response_400"
                    },
                    "401": {
                        "$ref": "#/components/responses/response_401"
                    },
                    "404": {
                        "$ref": "#/components/responses/response_404"
                    },
                    "422": {
                        "$ref": "#/components/responses/response_422"
                    },
                    "429": {
                        "$ref": "#/components/responses/response_429"
                    },
                    "500": {
                        "$ref": "#/components/responses/response_500"
                    },
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "schema": "UpdateItem",
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/response_schema"
                                        },
                                        {
                                            "properties": {
                                                "data": {
                                                    "properties": {
                                                        "phoneNumber": {
                                                            "description": "用户绑定的手机号（国外手机号会有区号）",
                                                            "type": "string"
                                                        },
                                                        "purePhoneNumber": {
                                                            "description": "没有区号的手机号",
                                                            "type": "string"
                                                        },
                                                        "countryCode": {
                                                            "description": "区号",
                                                            "type": "string"
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    ]
                                }
                            }
                        }
                    }
                },
                "deprecated": false,
                "security": [
                    {
                        "oauth2": [
                            "scope"
                        ]
                    }
                ]
            }
        },
        "/api/v1/wx-applet/wx-acode-unlimit": {
            "post": {
                "tags": [
                    "微信小程序"
                ],
                "summary": "",
                "description": "获取小程序二维码",
                "operationId": "获取小程序二维码，适用于需要的码数量极多的业务场景。通过该接口生成的小程序码，永久有效，数量暂无限制。",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "scene"
                                ],
                                "properties": {
                                    "scene": {
                                        "description": "最大32个可见字符，只支持数字，大小写英文以及部分特殊字符：!#$&'()*+,/:;=?@-._~，其它字符请自行编码为合法字符（因不支持%，中文无法使用 urlencode 处理，请使用其他编码方式）",
                                        "type": "string"
                                    },
                                    "page": {
                                        "description": "必须是已经发布的小程序存在的页面（否则报错），例如 pages/index/index, 根路径前不要填加 /,不能携带参数（参数请放在scene字段里），如果不填写这个字段，默认跳主页面",
                                        "type": "string"
                                    },
                                    "width": {
                                        "description": "二维码的宽度，单位 px，最小 280px，最大 1280px",
                                        "type": "string"
                                    },
                                    "auto_color": {
                                        "description": "自动配置线条颜色，如果颜色依然是黑色，则说明不建议配置主色调，默认 false",
                                        "type": "string"
                                    },
                                    "line_color": {
                                        "description": "auto_color 为 false 时生效，使用 rgb 设置颜色 例如 {r:xxx,g:xxx,b:xxx} 十进制表示",
                                        "type": "string"
                                    },
                                    "is_hyaline": {
                                        "description": "是否需要透明底色，为 true 时，生成透明底色的小程序",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "400": {
                        "$ref": "#/components/responses/response_400"
                    },
                    "401": {
                        "$ref": "#/components/responses/response_401"
                    },
                    "404": {
                        "$ref": "#/components/responses/response_404"
                    },
                    "422": {
                        "$ref": "#/components/responses/response_422"
                    },
                    "429": {
                        "$ref": "#/components/responses/response_429"
                    },
                    "500": {
                        "$ref": "#/components/responses/response_500"
                    },
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "schema": "UpdateItem",
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/response_schema"
                                        },
                                        {
                                            "properties": {
                                                "data": {
                                                    "properties": {
                                                        "phoneNumber": {
                                                            "description": "用户绑定的手机号（国外手机号会有区号）",
                                                            "type": "string"
                                                        },
                                                        "purePhoneNumber": {
                                                            "description": "没有区号的手机号",
                                                            "type": "string"
                                                        },
                                                        "countryCode": {
                                                            "description": "区号",
                                                            "type": "string"
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    ]
                                }
                            }
                        }
                    }
                },
                "deprecated": false,
                "security": [
                    {
                        "oauth2": [
                            "scope"
                        ]
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "response_schema": {
                "properties": {
                    "status": {
                        "description": "状态: [T->成功, F->失败]",
                        "type": "string"
                    },
                    "message": {
                        "description": "接口消息",
                        "type": "string"
                    },
                    "code": {
                        "description": "错误码",
                        "type": "string"
                    },
                    "data": {
                        "description": "数据",
                        "type": "object"
                    }
                },
                "type": "object"
            },
...
        },
        "responses": {
...
            "response_400": {
                "description": "请求错误",
                "content": {
                    "application/json": {
                        "schema": {
                            "schema": "UpdateItem",
                            "allOf": [
                                {
                                    "$ref": "#/components/schemas/response_schema"
                                }
                            ]
                        }
                    }
                }
            },
            "response_401": {
                "description": "用户未授权",
                "content": {
                    "application/json": {
                        "schema": {
                            "schema": "UpdateItem",
                            "allOf": [
                                {
                                    "$ref": "#/components/schemas/response_schema"
                                }
                            ]
                        }
                    }
                }
            },
            "response_404": {
                "description": "页面不存在",
                "content": {
                    "application/json": {
                        "schema": {
                            "schema": "UpdateItem",
                            "allOf": [
                                {
                                    "$ref": "#/components/schemas/response_schema"
                                }
                            ]
                        }
                    }
                }
            },
            "response_422": {
                "description": "Unprocesable entity - [POST/PUT/PATCH] 当创建一个对象时，发生一个验证错误。",
                "content": {
                    "application/json": {
                        "schema": {
                            "schema": "UpdateItem",
                            "allOf": [
                                {
                                    "$ref": "#/components/schemas/response_schema"
                                }
                            ]
                        }
                    }
                }
            },
            "response_429": {
                "description": "Too Many Requests - [*] : 用户在给定的时间内发送了太多的请求",
                "content": {
                    "application/json": {
                        "schema": {
                            "schema": "UpdateItem",
                            "allOf": [
                                {
                                    "$ref": "#/components/schemas/response_schema"
                                }
                            ]
                        }
                    }
                }
            },
            "response_500": {
                "description": "INTERNAL SERVER ERROR - [*]：服务器发生错误，用户将无法判断发出的请求是否成功",
                "content": {
                    "application/json": {
                        "schema": {
                            "schema": "UpdateItem",
                            "allOf": [
                                {
                                    "$ref": "#/components/schemas/response_schema"
                                }
                            ]
                        }
                    }
                }
            }
        },
...
        "securitySchemes": {
            "Password Based": {
                "type": "oauth2",
                "description": "Use a global client_id / client_secret and your username / password combo to obtain a token",
                "in": "header",
                "flows": {
                    "password": {
                        "authorizationUrl": "/oauth/authorize",
                        "tokenUrl": "/oauth/token",
                        "refreshUrl": "/oauth/token/refresh",
                        "scopes": []
                    }
                },
                "scheme": "https"
            }
        }
    },
...
}